refactor(spark): Extract and compile generator
